Dance around a serious downcasting problem
Right now, smt_sorts coming back into python aren't being converted to
the appropriate object reference. I know that it's available, because
the object wrapper class holds such a reference, and I can see b.p
initializing it via gdb. But, when calling into python this doesn't seem
to work.

It could be that we need to manually downcast smt_sort to the
specifically wrapped python type, to let python know that it can just
suck a python reference out of it. But, that jsut leads to "no by value
converter registered". This is a connundrum.