Perform special dance to move smt_convts to python
When calling from smt_ast's, we have to downcast the convt object that
was passed in too. However, because smt_convt_wrapper and
smt_ast_wrapper are mutually dependent, we can't just have each call
each other, we ahve to declare one and put it's definition further down
the file. Ugly, but works, and is a legitimate solution to a legitimate
problem.