Allow lambda's to iterate over fields of types
In the same way as you can pass a delegate down to all expr2tc fields of
an expr, this should allow exactly the same to be done for types.

Reason for requiring this: we need to whip all pointer types out of
certain type2tc's in smt_convt, and this is the most effective way to do
it.