Allow for non-local-storage rvalue conversion
That is; allow the from-python converters to avoid storing values in the
rvalue storage allocated by b.p. Instead, we just return the pointer
within a container. This is implemented for irep2tc_to_irep2t, which
*might* break our ownership model a little. It's potentially possible
that you could pass it into a shared_ptr that took ownership of it,
which would be bad/wrong. However, given some problems being faced, it's
necessary.

Specifically, b.p doesn't seem to wish to pass irep2tc's into
expr2t::operator==, so we have to mangle rvalues ourselves. Ugh.