Implement some extra irep operators
That is; allow expr2t-ref operator=='s to be called. Right now it
appears what's happening is that b.p won't pass irep2tc's into that
method, and instead returns a NotImplemented value, which doesn't cause
an error, just a false comparison.

Overload __eq__ with the correct kind of comparison operator.a This
seems to fix it.