Refactor SMT address-of calculations for structs
Essentially, if a composite type got fed into convert_addr_of, then the
first layer of index / member expressions gets handled, but some of the
offset data gets dropped. Which is massively wrong.

As it happens we already have sufficient functionality to create the
correct pointer arith expressions, it just didn't exist when the current
SMT backend was constructed, so I didn't use it. Sigh.