Fix numerous recursion failings
This fixes two locations that throw around level0 variables without
considering that their context dictates that actually, a specific l1
variable is being delt with. In particular: function return values need
to be assigned to the l1 variable of the caller (breaks on recursion),
and the phi function needs to create assignments between the l1 data
objects that have changed (breaks on recursion).

This introduces a static renaming::level2t method that renames a l0
symbol to a specific name_record's level1. This is relevant as there are
use cases where we have a handle on what l1 name we want to use, but
maybe no level2t object is appropriate.