Fix memory leak check when malloc can return NULL
If malloc returns null, then various programs out there correctly
observe this and don't attempt to free a null pointer. However, as far
as ESBMC is concerned that allocation still happened, it's just that
malloc returned null, and memory leak detection fires.

Fix this by adjusting the guard of an allocated hunk of memory so that
it doesn't register as being allocated if malloc returns null.