Revert "Avoid printing misleading dynamic_value merges"
This reverts commit 1bce01f71e671bbdb2343c01a59b06dc05569682.

On second thoughts: it's actually not the case that the current_number
== 0 means that we're merging an unallocating path. A feasible cause of
this is that the allocation happened before branching but without
assignment (leaving current_number == 0), then an assignment happens on
the branch.

We don't actually have the capacity to distinguish "we just allocated
this" and "we never assigned to it" at this time.