Implement by-bitvector typecasts
We have a need to distinguish the value-conversion of floats<=>ints and
the bit-representation conversion of the same. This patch adds a new
irep (bitcast) that piggybacks on typecast. To test this, I've
implemented the relevant conversions for mathsat, but not Z3.

It's possible that in the future we'll want to distinguish certain
conversions for conceptual nicety: the real implication of this irep is
to say "advanced pointer stuff happened here".

This patch is related to #105, although it doesn't fix it as more
testing is needed, plus Z3 support.