Fix some more type trait irregularities
Looks like some ireps base data was merged because they were the same --
however there was a difference, in that some of these types took a type
argument, and others didn't. This is important when registering
constructors with python. Thus, convert not_data into a bool_1op data
holder, re-usable by a bunch of other things.