Remove special case for array equalities
In the previous revision of convert_ast the code skipped converting
equality asts, only to directly convert them later on. This is probably
a relic of having to handle the array flattener, but before
convert_assign was introduced.

This is actually knackering what I'm working on at the moment, so a good
time to kill this "feature"