Account for VLAs in byte sizes, not elems
The pointer interpretation expects the dynamic_size field to be filled
with a byte-measured size, but we're feeding an element. Mutliply with
the size of the array subtype to fix this.

This will break if the array subtype is of unknown size, but that's fine
because it'd be illegal C then (IMO).