Handle byte_extract's in get_value_set
There's a legitimate pattern where one is assigning bytes one by one
into a dynamically allocated array, which involves byte extracting from
a bitcasted pointer.

Note that while it's tempting to replace all of this with a "just
collect all the symbols" approach, one still needs to track the mode of
"collecting references" (of addrof) and "collecting values" (of
pointers). The current implementation sucks, but is necessary.