Fix a variety of problems identifying bitfields
We need to blast through symbolic types; we also need to quickly look
things up in both maps to see whether or not this type has bitfields, is
fixed, or has a fix.

And then we need to fix it in location for member expressions to be
formed correctly. I don't enjoy the fact that a bunch of symbolic types
are being expanded ahead of time, but that *should* (TM) not make a
difference to anything.