Implement bitfield updates
This allows for assigning through an extract irep: I was hoping we
wouldn't have to do this, but it's impossible to take account of whether
you're in the lhs or rhs of an assignment in the frontend, it violates
all abstractions. Thus, at symex time we'll need to interpret an
assignment into a bitfield.

The attraction of the technique though is that because everything is
stored as a blob, access into a bitfield by a character pointer (i.e.
memcpy) doesn't require any further work or special casing: the field is
only treated as a bitfield when the lexical level referred to a bitfield
part of it.