Fix concatenation of assigned bit extract
If there's no bottom part of the original bitvector to concatenate in,
then the whole bottom part of the resulting bitvector is made of the
update value. However in the current implementation, we don't handle
that case, we just fall through an if condition.

Fix this by assigning the (accumulated) bottom part of the result
bitvector to be the update value, in this circumstance.