Mash up in-struct-field memsets
This is unpleasent: draw together a series of dereference related things
to allow us to generate references to random fields inside structs that
we *might* be memsetting to zero.

Specifically, if you're memsetting an 8 byte block in a struct to zero,
we can usually get away with just building references to everything that
might be a uint64_t and assigning to that expression. But, if it's 15
bytes (and thus might be some arbitary struct) within _another_ struct,
we need to repeatedly dereference the pointer with different types, to
build different kinds of references to... different fields. And then
assign different zero-initialized values to each.

Super unpleasent; only invovled some minor hacks to get the right
information out of dereferencing, full story will have to go in real
documentation (TM).