Mash around the nondet-offs-in-struct memset
It wasn't working very well where we relied on a nondet offset
dereference of a pointer. Instead, after recursing through struct types,
store offets and types and use those to later construct references.

This is worryingly like the dynamic offset dereference stuff -- there
might be a call for genericifying that code in the future, but that time
is not now.