Account for invalid reference construction
I'd assumed here that the only circumstances where one byte-extracted
from a byte was when one was extracting a byte. However that's not
always the case: #223 has a case where a pointer is constructed from a
byte field. This isn't valid, and isn't actually executed, it's just a
possibility.

Thus, only optimise out byte_extract's when we're _truly_ reading a
byte, i.e. when a single byte is requested and the source is one byte.

fixes #223 but ./ldv-memsafety/memleaks_test12_false-valid-free.i has
bigger issues.