Add a timestep field to value set tracking
It occurs to me that we're spending a lot of time comparing items that
haven't actually changed: it would be fine to track when different
symbols were assigned, and only merge when there'd been an update. Start
to do this using the number of SSA assignments as a timestep key, as
that precisely correlates with Things Happening in the execution of the
program.

Requires zero timestep for the static analysis facility. Removes a
variety of calls to value_sett::assign from intrinsics, because they all
call symex_assign, which eventually calls that function anyway.