Merge-mangle in phi functions
That is: instead of repeatedly accessing the current names map to work
out whether or not two symbols should be merged, instead use an ordered
map and incrementally step through, making use of the order to avoid
un-necessary comparisions.

It's worth noting that this and similar work are a trade-off: if there
are many more reads than merging activity then the loss from not using
hashing is great. Horses for courses.