Marry clang classes to defeat bitfields
The current "adjustment" phase for clang C conversion is insufficient,
there seem to be a lot of places where expressions drop through the
cracks and don't get their bitfields fixed. To resolve this, only ever
read-in fixed bitfield structs.

This involves moving most of the bitfield code over to the
clang_c_convertert class, and then passing information from that to the
adjuster phase so that subsequent member expression adjustment has
enough information to operate. IMO, this is proportionate sharing of
metadata blah blah.

A small amount of dancing has to happen around the reading in of
constant struct expressions, but that's to be expected.