Allow mutation of fields in ireps from python
This has numerous problems associated with it: if you just do it to some
arbitary irep, you have no idea who else has a reference to it, and
you're screwed. Likewise, if the crc of the irep isn't invalidated when
you touch it, you're also screwed.

The only safe way to mutate ireps is by cloning it so that you know
there's new storage allocated, then iterate through each subexpr to
clone that away too, then perform your mutation, and then invalidate
crcs from the top. Hence, all these fieldnames are prefixed with the
estople "iknowwhatimdoing".

Or, y'know, you could just not do this in the first place.