Attempt to determine local variable liveness
When we dereference a pointer and potentially point at a non-malloc'd
variable, check to see whether or not it's still alive. If it's still
recorded in the l1 renaming map in one of the functions on the call
stack where it was allocated, then in a corresponding real program,
storage would be allocated, and the pointer would be valid.

If not, then were the pointer to point at this variable, this would be
an illegal access.

For example, 01_cbmc_Pointer20 fails, because it's horribly broken.