[filesystem] replace tmp_dir by proper abstraction tmp_path + add tmp_file
The race condition inherent to all methods of creating temporary files (or
directories) that separate path construction from obtaining the corresponding
resource can lead to abort() in the existing implementation
file_operations::get_unique_path(), even though it contains a loop exactly
to prevent that.

On POSIX there is mk{s,d}temp(3p) to get around this loop, however, for temp
directories on Windows it does not seem that this functionality can easily be
replicated.

We avoid the abort() case by moving the creation of file/dir into the loop,
which now only fails (with a boost::filesystem::filesystem_error being thrown),
when all names of the given pattern are already used. In case the above race
condition is triggered, the new implementation can still loop unboundedly.

Additionally, two new classes, tmp_path and the derived tmp_file, are
introduced which take care of removing the respective path from the file
system upon destruction, and, in case of the latter, also closes the
associated FILE. The purpose of this RAII style API is to encapsulate the
associated resource(s), freeing users of the finalization tasks.