[cmake LLVM] added CMake options to control bundling of clang headers
Added the new CMake variable CLANG_HEADERS_BUNDLED, which can take
(detect, esbmc, On, Off) defaulting to 'detect' and controls whether
bundling of headers should be enabled:
- detect: based on BUILD_STATIC
- esbmc: enabled and use headers from the source distribution
- On: enabled and use the headers in clang's resource-dir (see below),
      unless overriden by the new CMake variable OVERRIDE_CLANG_HEADER_DIR
- Off: disabled, hard-code path to clang's resource-dir and pass it via -I
       directly to libclang-cpp

The default 'detect' is different from previous behaviour ('esbmc') in that
non-static builds against the system's clang now default to using the system
clang's headers.

Clang's resource-dir is determined by compiling scripts/cmake/try_clang.cc
against the headers and libraries found in Clang_DIR (or LLVM_DIR, if unset),
extracting the value of the C-macro CLANG_RESOURCE_DIR used by the clang
tooling to derive it's own resource-dir:

  <https://clang.llvm.org/doxygen/Tooling_8cpp_source.html#l00462>

Additionally, when BUILD_STATIC and the final value of CLANG_HEADERS_BUNDLED
do not match, warnings about possible inconsistencies on updates are printed
by CMake. These are necessary in order to inform users about the dependency
between the headers and libclang.

See scripts/cmake/FindLLVM.cmake for details.

TODO: One of the cases introduced by this change is not implemented:
CLANG_HEADERS_BUNDLED=On and OVERRIDE_CLANG_HEADER_DIR is set. In this case,
ideally, we should bundle the headers in the given directory and re-generate
src/clang-c-frontend/clang_headers.cpp, but the infrastructure is not set up,
yet.