[testing-tool] avoid any chdir before invoking TOOL
This patch makes the testing_tool oblivious to the current working directory.

Generalizes the previous patch by not touching the TOOL parameter in any way
except splitting the string according to shell rules. This has the effect that
arbitrary runtime environments for esbmc can now be specified using the TOOL
paramater, e.g.,

  ../regression/testing_tool.py --tool
    "qemu-aarch64 -L /usr/aarch64-linux-gnu -E [...]
       ../build/src/esbmc/esbmc --fixedbv" [...]

The path to the regression suite is now relative to that of testing_tool.py,
which is determined at runtime. The file paths constructed for the command
line are now generated by prefixing each with the regression suite's path.

This has the additional effect that the current working directory no longer
influences the way the regression suites are found.