[cmake] rm special case when CMAKE_LIBRARY_ARCHITECTURE is set
This construction was supposed to produce a path like

	/usr/include/x86_64-linux-gnu

on multilib (also known as multiarch) systems in order to include headers
corresponding to the multilib "architecture" for which stuff is being built.

It has a few problems:

a) This setting is only meaningful for library search paths and was not
   intended to be used in such a way, cf. CMake docs.
b) The compiler is already doing "the right thing" on multilib systems
   depending on the particular target configuration (i.e. -m32 vs. -m64).
c) Enabling it unconditionally (see b) above) may lead to inclusion of a
   64-bit header in -m32 mode.

That's why we drop it.