[elibc] pthreads: simplify definitions for operational model
This removes all platform-specific definitions/declarations from the internal
model of pthreads. The goal is to unify the interpretation of user code across
all platforms/system. This patch therefore removes the Glibc-specific parts
as well as unnecessary fields from the defined types, as well as their
embedding into a union.

The second change concerns the global variable 'head' defined in pthread_lib.c:
It was exported as an external symbol before, which the regression test
nonz3/01_cbmc_Malloc18 did uncover, as there is a conflict with the symbol of
the same name declared there. Therefore, pthread_lib's 'head' now has internal
linkage.