[regression] add 2 failing and 1 working tests from #709
The two failing tests are those that use "cast to union", which before the
recent introduction of new symbols for compound literals had been handled
the same way: by a constant_*2t expression for the construct.

Thus, the compound literals case works as expected now, but the cast to
union still fails.