[clang-c-frontend] restore recursive call to help Boolector timings
In <https://github.com/esbmc/esbmc/pull/764#issuecomment-1166303300> the
recursive call to get_expr() was identified to restore Boolector timings
even though the resulting exprt of the call is now unused since we employ
Clang to compute the bit-field width for us.

Those side-effects causing slight GOTO reordering and massive slow-down
of Boolector on the SV-COMP benchmark

  c/ldv-validator-v0.6/linux-stable-c0cc359-104_1a-drivers--usb--serial--qcserial.ko-entry_point.cil.out.i

need to be identified and probably somehow removed.