[mode] factor out language-ID from the mode_table
Give the supported languages proper IDs so we can use the language information
down the line. Before, the language/frontend was identified by a free-form
string only, however, in #776 it was deemed necessary to assign a meaning to
options like `--overflow-check` that depend on the concrete language we're
verifying over.

An additional benefit is that the name of frontends now is independent of the
language they're translating, allowing us to freely assign and show them in the
log later on.

The `mode_table` itself for now retains its weird semantics wrt. the old
frontends.

Further, we get rid of some globally exported symbols and mark others `const`
like `buildidstring_buf` and `esbmc_version_string`, so they do not count
toward global "state" #755.