[symex] assume addr-space layout constraints when tracking new ptr instead of in smt_convt
The address space layout constraints make the assumption that all (alive,
fixes #775) dynamic objects have disjoint addresses. This was ensured in
smt_convt::finalize_pointer_chain() before, but the validity of objects
could not be ensured there, see #781. By assuming those conditions in
symex, the validity of objects can and now is encoded.

This also fixes a wrong type for the __ESBMC_alloc_size[obj] expression.