[util] move "merged context" functionality from namespacet to only user c_linkt
The namespacet class was an amalgamate of two separate functionalities:
Lookup or migrate symbols/types and (optionally) merge this functionality for
two different contexts. We now separate this second "merged context" from the
basic namespacet into a new class merged_namespacet. As the only user of these
merged namespaces is c_linkt, put the class definition there. Also, c_linkt
does not export or record its merged namespace anywhere, it only uses it in the
process of modifying the contexts it links.

This has the effect that a namespacet for the rest of ESBMC now only depends on
one single context, becoming basically a glorified wrapper for it. Makes
reasoning about namespaces throughout easier, as we can be sure there is only
one context.