[clang-c-converter] fix initialization of static storage compound literals
Earlier, the initialization order for those comp-lits was wrong, e.g.

  int vsi = (int){42};

in file scope got translated to

  vsi=main.c:7$compound-literal$1;
  main.c:7$compound-literal$1=42;

The reason was the unintuitive but necessary behaviour of `get_var()` which
takes care of variable declarations including the initializer:

  int x = x;

is allowed, so the RHS (in general, but not in the static case!) needs to be
able to access the symbol on the LHS.

This fix is by special-casing the static case, as the RHS there has to be a
constant expression and thus cannot refer to other symbols (with the exception
of compound literals). So first translate the RHS (pretending there is no
enclosing scope to accomodate

  void f() {
    static int x = (int){42};
  }

by temporarily clearing the `current_block` until the static decl is finished)
and then create the symbol for the LHS.