[cmake] add support for FetchContent of Catch2 v3.x
Summary: add v3.x support for Catch2 as the cheri-compressed-cap library wants
that one.

FetchContent is utterly broken [1]. As usual, it implements its own "dependency
management", which consists of the, quote,

> "first to declare, wins" approach

Not only does that break everything importing another package as then that
project could potentially override other packages down the line, no, it also
introduces even more state into the CMake files in form of sequence of
declarations. However, FetchContent is not the only part of the build system
that relies on the specific order of declarations. This results in projects
using FetchContent packages to solve the nice NP complete dependency
management problem manually in their CMakeLists.txt every time a dependency is
updated...

[1] <https://gitlab.kitware.com/cmake/cmake/-/issues/22904>