[smtlib] handle death of ext process via exception
This way the smtlib error message is given to the user.

We handle this by setting the SIGPIPE handler to SIG_IGN, such that
fwrite(), fflush(), etc. will return an error when the child process
dies and they will set errno to EPIPE. We now handle this condition
by reading all the remaining characters from the input pipe and
throwing an exception containing this string. stderr is still shared
with that process such that any message printed there will also appear
in ESBMC's error output.

Typically, SMT solvers are not supposed to die when they encounter an
error, but a notable exception to this rule is Yices 2.