[smt-conv] it2t's type can differ from either side's
This addresses discrepancies in the translated SMT solver's types
between that of the if2t and those of either of its side's. In
regression/k-induction-parallel/invert_string for instance, the Z3
and CVC4 backends aborted due to such a construct where the if2t's
array size is a small constant but either side's one is a symbol of
a larger type.