[symex] rename_type() should also recurse into operands
`goto_symex_statet::rename_type()` did only consider the expression's type and
its subtypes and not sub-expressions as well. This causes the sub-expressions
for instance in the if2t `cond ? array1 : array2`, to not be considered,
triggering issues with the Z3 and CVC4 backends due to mismatching bv-widths
for array sizes. Something that's supposedly fixed up by the internal
fixup_renamed_type() for other kinds of expressions, but for some reason that's
only invoked for symbols, not ifs, and for renaming exprs, not types.
So the the recursion is put into the public `rename_type()` interface instead.

The other backends using array_conv have not been affected because they simply
ignore the if2t's type annotation and go with that of the true_value sub-expr.