[regression] mark KNOWNBUG due to changes of is_correctly_aligned in dereference
The computation of is_correctly_aligned was updated to correctly take into
account the offset. For the following two tests, the offset is not constant
but (N-1)*sizeof(float). In this dynamic-offset case we cannot (easily)
ensure that the dereference uses the correct alignment and fall back to
byte-extraction. This, however, is not supported for --ir, thus we abort.

It affects the three existing regression tests:
- regression/esbmc-unix2/20_flex
- regression/cbmc/01_cbmc_Fixedbv23
- regression/cbmc/01_cbmc_Fixedbv24

Also add copies of these tests without the --ir. However, the copies of the
last two fail, too. This time it is independent of the change to determining
correct alignment. There may be another bug. See issue #1215.