[clang-c] convert: the definition of the type might come later than first use under pointers
Avoids throwing an inf_sized_array_excp for incomplete_types for which
Clang does not have a definition where they are first used (as a pointer
for instance). ns.follow() on symbol_type2t would then via migrate
construct an infinite array type, which is clearly not what's meant by
using a forward-declared type.

Fixes:
- sv-benchmarks/c/ldv-linux-4.0-rc1-mav/linux-4.0-rc1---sound--pci--ice1712--snd-ice1724.ko.cil.i
- sv-benchmarks/c/ldv-linux-4.2-rc1/linux-4.2-rc1.tar.xz-43_2a-drivers--net--ethernet--mellanox--mlx4--mlx4_en.ko-entry_point.cil.out.i

  Here, the exception got thrown in add_padding() on a struct member's
  type symbol since the context did only contain an incomplete_struct
  even though there is a complete definition in the file:
  `tag-struct mlx4_cq` as first member of `struct mlx4_en_cq`.

  The incomplete_struct was put into the context during definition of
  the type `tag-struct __anonstruct_tasklet_ctx_330`, which
  a) is defined before `tag-struct mlx4_cq`, so Clang doesn't have a
     complete type definition yet at that point, and
  b) contains a pointer to `tag-struct mlx4_cq` type.

The fix is by
- continuing to try to construct the complete type even if it is in
  the context, subject to irept::incomplete() being set, and
- using the .incomplete() flag to stop the infinite recursion in case
  we're already constructing this type up the recursion stack.

In conjunction with using symbol_type anywhere else to refer to the
actual type sitting in the context, this enables us to finally get
the correct definition for forward-declared types that are used under
pointers.