[regression] relax check
The regex checks whether some RHS in the counter-example matches another one,
which doesn't necessarily have to be the case. Since its a reduced example
of #447 which I think has a correct regex checking the same condition on
different code, instead of opening an issue to find a matching regex lets
relax the test performed here.