[irep2] remove constructors of *2tc from a) raw pointers and b) as conversions from the base2tc
The purpose is to make the C++ syntax conform more to the default expectations
where

  Type v(expr);

constructs a *new* variable out of 'expr'. It is not a typecast, but a
constructor. Thus, this patch sets all the 'always_construct' type traits
to 'true'. There is no more confusion whether

  not2tc(expr);

is a type-cast or a construction of a new shared-ptr to a not2t constructed
from 'expr'. It is the latter.

This brings along some difficulties when trying to build something like a
symbol2tc from an expr2tc. It's basically not possible anymore. But that's OK
since the only time you'd want to do that is when manipulating the underlying
symbol2t. So you're now expected pour in some work for that, because you
shouldn't actually do that! irep2 is supposed to be immutable.

So now the 'always_construct' type trait is a bit useless.