[irep2] rm unused and wrong expr2t methods num_nodes() and depth()
They were wrong because they didn't take into account that irep2
forms a DAG. The notion of #nodes does exist for DAGs but it is
supposed to only count *unique* reachable nodes, I suppose. While
depth is a concept that's inherent to trees only. There is tree-width
for undirected graphs, but it's a bit different...