[irep2] rework comparison on irep2tc + more docs
The set of comparison operators is complete now (for C++17) on
irep_container (aka expr2tc and type2tc).

Also add a check whether the addresses of the pointees are equal.
This about halves the running time of the somewhat
comparison-intensive

  regression/esbmc/17_apache

for a Debug build and reduces its symex time to ~60% for a Release build.