[cmake] use check_linker_flag() instead of compiler id
The linker used by the compiler is not determined by its name.
For instance, on typical Linux distributions the Clang compiler defaults to
GNU ld (aka ld.bfd) while on macOS Clang seems uses Apple's ld64, which
accepts options similar to ld.lld, but different from ld.bfd or ld.gold.

So instead we test all the variants that are known to us.