[smt-conv] get(): only recurse if there is a sub-expr
The tuple-node-flattener does not support getting arrays inside tuples right
now, so it generates a nil expr for the cases tested by the 2 new regression
tests. Recursing into nil exprs however leads to segfaults.

Fixes #1408. But it does not fix the underlying issue: When the
structures/arrays are nested deeper, this will come back to bite us.

A proper fix could be to return the "have_all" boolean from get() (and also
from the *-flatteners' versions of it) and properly use that information.