[libm] move Windows #defines from implementation to overridden headers
The purpose of these #defines is to prevent the Windows "ucrt" headers from
defining inline versions of the functions ESBMC provides operationals for.

Generally, the to-be-verified program should be translated in the same setting
as the c2goto libraries in order for types to be defined consistently.
Therefore move the #defines to the c2goto/headers, which override the system's.