[c-expr2string] (un)signedbv constant: format LLONG_MIN and things > LLONG_MAX according to strict literal rules
Background: the constant expression -9223372036854775808 is actually not
supported by C, because the unary - and the rest are not handled together.
That means that the literal 9223372036854775808 will be interpreted as
an unsigned long long int and the (implicit) cast to signed long long int
will be undefined (because it will touch the sign bit).

The second change done here is to output any constant > LLONG_MAX in hex
format. This is to please CPAchecker, a tool on a mission to banish large
numbers...