[symex] symex_input: use source location of intrinsic for assignments
In symex_step() the PC is incremented before executing the intrinsic,
this leads to assignments made by intrinsics to have the wrong source
location. Here we fix the symex_input() family of intrinsics to use
the correct source location.

For counter-examples and witnesses to also print these assignments,
additionally to the other conditions on goto-steps allow function
call steps to be interpreted as assignments - if we do that internally.

This fixes a warning printed by CPAchecker about the assigned symbol
not being on the source line indicated in the witness file for (at least)
SV-COMP benchmarks no-overflow
1. c/Juliet_Test/CWE190_Integer_Overflow__int64_t_fscanf_square_21_bad_abs.i
2. c/Juliet_Test/CWE190_Integer_Overflow__int64_t_fscanf_add_08_bad.i

Unfortunately, CPAchecker still doesn't want to find the property violation,
but with --compact-trace this now enables UAutomizer to validate the second
one above.

Related: #1470, though we still can't mark it fixed because of --compact-trace.
It's really unclear why --compact-trace makes a difference for item 2. above,
I can't see anything in the larger counter-example that would be of help for
validating the witness...