[clang-cpp] Advance iterator even when identifier is empty
Previously, when an identifier was empty, we would skip the loop iteration
by `continue`-ing it. This however had the effect that the iterator `it1`
would not be advanced, because that statement was at the end of the loop
and we just skipped it.
Move the advancement of the iterator such that it is always executed once
per loop iteration regardless of whether we might skip parts of the current
iteration.