[clang-c] language: set Clang's resource dir with correct option
The resource-dir/include directory will be put into the header search paths
after all the -isystem ones. Though there is -isystem-after, unfortunately,
only the cross-Windows frontend of Clang supports it and all others don't
"use" that option (i.e., they ignore it).

Giving the directory via -resource-dir is the correct way to go here, but
it requires the "include" directory below it to contain all the expected
Clang headers.