Associate need for vptr/vbot_ptr with type instead of converter instance
Associating the need for vptr/vbotr_ptr with converter would mean that constructors implemented outside the declaring class
would not have their vptrs set, because by the time the converter reaches the implementation
the has_vptr_component would already have been reset to `false`, because it is always reset when a class
has been finished parsing.
Therefore, associate the need directly with the type.