[clang-c/cpp] Don't specialcase CLike CXXRecords
This breaks tests related to pthreads, because we precompile pthreads with a C compiler.
When running the actual test, we have pthread methods that have been compiled by a C compiler
but also pthreads methods that have been compiled by a C++ compiler. Since the layout of
types is different between C and C++ this breaks.
Ignore this for now.