[clang] implement __builtin_{memcpy|memmove}
Implement __builtin_{memcpy|memmove} by defining them to be equal to __esbmc_builtin_{memcpy|memmove}.
We have to do this, because clang forbids us from directly implementing a function called __builtin_{memcpy|memmove}.
__esbmc_builtin_{memcpy|memmove} then simply delegates to {memcpy|memmove}.
We add the functions to the internal additions header, because they are builtins and should be available without
having to import any headers. (The implementation of string.c is available implicitly).