[simplify, TCs] fix nan infinite recursion in simplify
NaN != NaN is always true, but this defeats the purpose of the check
The assumption is that `src != source_value` means that we were able to simplify
something and return a simplified expression.
In case of NaN, we would continuously return "simplified" expressions, because
NaN != NaN would always be true.