[clang-cpp] use more robust method to get destructors
The old method to get destructors would perform a lookup using the name of the class
being destructed.
This caused problems, because the destructor would be named "~axy" while the fully qualified name
of the class would be "axy@2234".
We now use the more robust method that searches the type's components for a method
that is marked as destructor.

It's possible to change back the fully qualified name to "axy", however we would have to
decouple the way we create ids for structs.
Previously, "struct axy{}" would get name/tag "axy" and id "tag-axy".
This immediately causes problems, when we have multiple "struct axy{}" that are defined in different
functions, because they would all get the same id.
Just generating a unique id - and keeping the name/tag simple - was also an idea, but currently
ESBMC relies on the fact that "tag-" + $tag returns the id of a struct.
So having an id like "c:@foo@bar@axy" would not work, because it would break ESBMC's expectation
on how to go from a name/tag to the associated id.
I deemed changing all the various places that rely on this assumption too much work for now.